WIP - CI uses self build valgrind
apparently the valgrind that's included in Bionic (v3.13) produces some
warnings that we can't easily avoid without risking hiding real issues.
both issues happen in tests that use DEBUG RESTART.
one is that it sometimes produces empty reports (probably a result of
using exec), and the other is that it warns about closing fds (the loop
we have in restartServer), but we can't find any way to skip just these.
see #7199